Control the Ethereum user_state with simple actors
Introduce a notion of SimpleActor with synchronous write transactions
and asynchronous read transactions, and have a lockable table to
associate an actor to each user.

Enhance the contract creation: create a new one if there is no address
in the DB, but fail loudly if there is an address in the DB but it is invalid.

In the Makefile, replace - by _ in target names to fit file names.

Rename the api_scgi executable to side_chain_client.

Fix gitlab-ci script