Add join to MonadS, rename our Lwt to Lwter
Add the join operator to the MonadS signature. Join collapses two
levels of monadic wrapping into one, and we are using it in
ethereum_user. However, this standard name (used in Haskell and other
OCaml monadic libraries) conflicts with the existing Lwt.join
primitive, so we can't claim anymore that our Lwt module is a
conservative extension of the existing Lwt module. Thus, we rename
ours to Lwter (bad pun between "more Lwt" and "Later", also rightfully
suggesting that this is all a kluge).

Make tiny progress on the client side. Notably, memoize the making or
loading of TransactionTracker objects, so we never have two concurrent
handlers racing to handle the same transaction.