Subtle DB fix and improvement
Don't hold a transaction merely when there is another one
running, only when there's one running but we're ready for commit.
The former behavior creates too much contention, which is bad for
performance, though it's good for detecting deadlocks, so we may want
to enable it during some tests, although we then should probably
enable other transactions as soon as "the" transaction has returned,
though not if already ready && triggered.

Also add committing and committing_async primitives.

Also add a note on making DigestValue references weak.