Incomplete attempts at improving ethereum_user
Implement NonceTracker, but it leads to deadlock because messages are
sent out of order to geth, that drops the early ones, then deadlocks.

Implement retrying in a more uniform way, but it leads to race
conditions and many multiple copies of a transaction being sent.

Start a long discussion of things to be done in TODO.