Ethereum_user: Always use PreTransaction
Always use the PreTransaction abstraction, and never build
transactions directly. Thus, we can enforce a policy of only
one transaction at a time, and, in the future, use a batching
contract.

Also, use eth_estimateGas by default now that we have
pre-transactions so we don't have to guess so hard.