Create tests to check for TransactionRejected
Created tests `contract-success-on-Ethereum-testnet` and
`contract-failure-on-Ethereum-testnet` in `ethereum_user.ml`. The
failure test uses a bogus gas_limit to trigger a TransactionReceipt
with a status of 0, throwing `TransactionRejected`, which then returns
finally with `TransactionFailed`.

I left the test `create-contract-on-Ethereum-testnet` even though it
is redundant now. I planned on removing it when I was done.