Generalize contract configuration, debug batching
Add a notion of per-address batching contract (in practice, only used
for Croesus for now), debug the EVM code for it, add OCaml support.
Include a test for the batch transfer contract.
Shuffle some code away from Ethereum_user into Batch.

Generalize a notion of contract configuration, saved either as file or
in the database. Use it instead of previous ad hoc mechanism for "the"
side-chain contract.

Rename ~value_send to simply ~value.